INFINITE PROCEDURAL GENERATION!!!!!!!!!!!!!!! The world now generates infinitely.
Added ProcGenerator, which functions almost identically to the LevelGen, except it's list of Sections only takes 1 of each section, and randomly chooses from the list for which section to generate next.

Corners now check to see if the player has reached it, when the player does, it destroys the section the player just completed and adds generates the section ahead of the current one. This means there's only ever 2 sections generated at a time)

When the player dies, the map is regenerated as usual, however as the Map only keeps track of 2 sections at a time, and completed sections are destroyed, the player respawns at the section they died in, functioning as a surprisingly effective respawn system.